// src/models/training-master.model.ts
export interface TrainingMaster {

  topic: string;
  activity_type_id: number;
  participants_type_id: number;
  level_of_participants: string;
  scheme_id: number;
  no_of_batch: number;
  expected_duration_of_training: number;
  no_of_expected_trainees: number;
  mode_of_training_id: number;
  category_of_activity_id: number;
  topic_description: string;
  financial_year: string;
  createdby: number;
  district: string;
  subdivision: string;
  range: string;

}